New path_dropbox_2dii()
Relates to #1 and #3, and closes #2.

This function is identical to `DROPBOX.PATH()` but has a better
name and documentation. This function aims to help people move
their projects out of 2dii's Dropbox flder -- which has a
problematic path.

DROPBOX.PATH() is now redundant and should be deprecated (#3).

Thanks @cjyetman and @pranavpandya84.